% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquet_readr_list.R
\name{parquet_readr_list}
\alias{parquet_readr_list}
\title{parquet_readr_list}
\usage{
parquet_readr_list(path, verbose = FALSE)
}
\arguments{
\item{path}{The path to the directory where the file is stored.}

\item{verbose}{For debugging the path string.}
}
\value{
A list of dataframes.
}
\description{
parquet_readr_list
}
