% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquet_readr.R
\name{parquet_readr}
\alias{parquet_readr}
\title{Title}
\usage{
parquet_readr(path, verbose = FALSE)
}
\arguments{
\item{path}{The path to the directory where the file is stored.}

\item{verbose}{Used for debugging the path string.}
}
\value{
A single dataframe of your multi-part parquet.
}
\description{
Title
}
\examples{
\dontrun{
# parquet_readr(path=path/to/your/parquet.parquet)

}

}
